# slack-mcp-server-api/openapi.yaml
openapi: 3.1.0
info:
  title: Slack MCP Server API
  description: A comprehensive API for searching and retrieving Slack messages, threads,
    and user information
  version: 1.0.0
paths:
  /messages/search:
    get:
      tags:
      - messages
      summary: Search Slack messages
      description: "Search for messages across Slack channels using various filters.\n\
        \    Supports text search, user filtering, date ranges, and channel-specific\
        \ searches.\n    Requires a valid Slack bot token with search:read scope."
      operationId: search_messages
      security:
      - HTTPBearer: []
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          description: Search query text
          examples:
          - important meeting
          title: Query
        description: Search query text
      - name: channel
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Specific channel ID to search in
          examples:
          - C1234567890
          title: Channel
        description: Specific channel ID to search in
      - name: user
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by user ID
          examples:
          - U1234567890
          title: User
        description: Filter by user ID
      - name: after
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Search messages after this date
          title: After
        description: Search messages after this date
      - name: before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Search messages before this date
          title: Before
        description: Search messages before this date
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of results to return
          default: 20
          title: Limit
        description: Maximum number of results to return
      responses:
        '200':
          description: Search results with matching messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSearchResponse'
        '401':
          description: Authentication failed - invalid token
        '403':
          description: Insufficient permissions
        '429':
          description: Rate limit exceeded
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-gram:
        name: search_slack_messages
        summary: Search for messages in Slack channels
        description: "\n<context>\nThis tool searches for messages across Slack channels\
          \ with comprehensive filtering options.\nIt can search by text content,\
          \ specific users, date ranges, and within specific channels.\nPerfect for\
          \ finding important conversations, tracking mentions, or analyzing communication\
          \ patterns.\n</context>\n<prerequisites>\n- If you have a channel name instead\
          \ of ID, use the list_channels tool first to get the channel ID\n- If you\
          \ have a username instead of user ID, use the get_user tool first to get\
          \ the user ID  \n- Ensure the bot has appropriate permissions to search\
          \ the target channels\n</prerequisites>"
        responseFilterType: jq
  /messages/{channel_id}/{message_ts}/thread:
    get:
      tags:
      - threads
      summary: Get message thread
      description: "Retrieve a complete thread including the parent message and all\
        \ replies.\n    Useful for getting conversation context around a specific\
        \ message.\n    Requires a valid Slack bot token with channels:history scope\
        \ and access to the channel."
      operationId: get_thread
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          description: Channel ID containing the message
          title: Channel Id
        description: Channel ID containing the message
      - name: message_ts
        in: path
        required: true
        schema:
          type: string
          description: Timestamp of the parent message
          title: Message Ts
        description: Timestamp of the parent message
      responses:
        '200':
          description: Thread information with parent message and replies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '401':
          description: Authentication failed
        '404':
          description: Message or thread not found
        '403':
          description: Access denied to channel
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-gram:
        name: get_slack_thread
        summary: Retrieve complete message thread with replies
        description: '

          <context>

          This tool fetches an entire conversation thread, including the original
          message

          and all replies. Essential for understanding the full context of discussions

          and following conversation flows in Slack.

          </context>

          <prerequisites>

          - You need both the channel ID and the timestamp of the parent message

          - The message timestamp should be the thread_ts value from search results

          - Ensure the bot has access to the channel containing the thread

          </prerequisites>'
        responseFilterType: jq
  /users/{user_id}:
    get:
      tags:
      - users
      summary: Get user information
      description: "Retrieve detailed information about a Slack user including profile\
        \ details, \n    real name, and other metadata. Requires a valid Slack bot\
        \ token with users:read scope."
      operationId: get_user
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          description: Slack user ID
          examples:
          - U1234567890
          title: User Id
        description: Slack user ID
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlackUser'
        '401':
          description: Authentication failed
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-gram:
        name: get_slack_user
        summary: Get detailed information about a Slack user
        description: '

          <context>

          This tool retrieves comprehensive user information from Slack, including

          profile details, contact information, and user status. Perfect for understanding

          message authorship and getting user context.

          </context>

          <prerequisites>

          - You need the user ID (starts with U) not the username

          - If you only have a username or display name, search for it first

          - User information availability depends on workspace privacy settings

          </prerequisites>'
        responseFilterType: jq
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MessageSearchResponse:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/SlackMessage'
          type: array
          title: Messages
          description: List of matching messages
        total:
          type: integer
          title: Total
          description: Total number of results
        has_more:
          type: boolean
          title: Has More
          description: Whether more results are available
      type: object
      required:
      - messages
      - total
      - has_more
      title: MessageSearchResponse
      description: Response model for message search results
...
    SlackMessage:
      properties:
        ts:
          type: string
          title: Ts
          description: Message timestamp
        text:
          type: string
          title: Text
          description: Message text content
        user:
          type: string
          title: User
          description: User ID who sent the message
        channel:
          type: string
          title: Channel
          description: Channel ID where message was sent
        thread_ts:
          anyOf:
          - type: string
          - type: 'null'
          title: Thread Ts
          description: Thread timestamp if message is part of a thread
        reply_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reply Count
          description: Number of replies in thread
        reactions:
          anyOf:
          - items:
              type: object
            type: array
          - type: 'null'
          title: Reactions
          description: Message reactions
      type: object
      required:
      - ts
      - text
      - user
      - channel
      title: SlackMessage
      description: Model representing a Slack message
    SlackUser:
      properties:
        id:
          type: string
          title: Id
          description: User ID
        name:
          type: string
          title: Name
          description: Username
        real_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Real Name
          description: Real name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
          description: Email address
        is_bot:
          type: boolean
          title: Is Bot
          description: Whether user is a bot
        profile:
          anyOf:
          - type: object
          - type: 'null'
          title: Profile
          description: User profile information
      type: object
      required:
      - id
      - name
      - is_bot
      title: SlackUser
      description: Model representing a Slack user
    ThreadResponse:
      properties:
        parent_message:
          $ref: '#/components/schemas/SlackMessage'
          description: Parent message of the thread
        replies:
          items:
            $ref: '#/components/schemas/SlackMessage'
          type: array
          title: Replies
          description: Thread replies
        reply_count:
          type: integer
          title: Reply Count
          description: Total number of replies
      type: object
      required:
      - parent_message
      - replies
      - reply_count
      title: ThreadResponse
      description: Response model for thread information
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...